requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.9.0
pymongo==4.7.2
python-dotenv==1.0.1
sentence-transformers==5.2.0
torch==2.9.1
transformers==4.57.3
scikit-learn==1.8.0
numpy==2.3.5
